+++ requester/handler.hpp
@@ -146,7 +146,7 @@ class Handler
         if (this->handlers.contains(key))
         {
             error("The eid:InstanceID {EID}:{IID} is using.", "EID",
-                  (unsigned)key.eid, "IID", (unsigned)key.instanceId);
+                  key.eid, "IID",key.instanceId);
             auto& [request, responseHandler,
                    timerInstance] = this->handlers[key];
             request->stop();
@@ -154,7 +154,7 @@ class Handler
             if (rc)
             {
                 error("Failed to stop the instance ID expiry timer. RC = {RC}",
-                      "RC", static_cast<int>(rc));
+                      "RC", rc);
             }
             // Call response handler with an empty response to indicate no
             // response
@@ -252,7 +252,7 @@ class Handler
         if (handlers.contains(key))
         {
             error("The eid:InstanceID {EID}:{IID} is using.", "EID",
-                  (unsigned)eid, "IID", (unsigned)instanceId);
+                  eid, "IID", instanceId);
             return PLDM_ERROR;
         }
 
@@ -298,7 +298,7 @@ class Handler
             if (rc)
             {
                 error("Failed to stop the instance ID expiry timer. RC = {RC}",
-                      "RC", static_cast<int>(rc));
+                      "RC", rc);
             }
             responseHandler(eid, response, respMsgLen);
             instanceIdDb.free(key.eid, key.instanceId);
