+++ utilities/requester/set_state_effecter.cpp
@@ -54,7 +54,7 @@ int main(int argc, char** argv)
     }
     pldm_msg* response = reinterpret_cast<pldm_msg*>(responseMsg);
     info("Done. PLDM RC = {RC}", "RC", lg2::hex,
-         static_cast<uint16_t>(response->payload[0]));
+         response->payload[0]);
     free(responseMsg);
 
     return 0;
